add observable error property to codecellviewmodel
