+++ src/vs/workbench/contrib/notebook/browser/contrib/cellDiagnostics/DiagnosticCellStatusBarContrib.ts
@@ -8,12 +8,13 @@ import { localize } from 'vs/nls';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
 import { OPEN_CELL_FAILURE_ACTIONS_COMMAND_ID } from 'vs/workbench/contrib/notebook/browser/contrib/cellCommands/cellCommands';
+import { CellDiagnostics } from 'vs/workbench/contrib/notebook/browser/contrib/cellDiagnostics/cellDiagnosticEditorContrib';
 import { NotebookStatusBarController } from 'vs/workbench/contrib/notebook/browser/contrib/cellStatusBar/executionStatusBarItemController';
-import { INotebookEditor, INotebookEditorContribution, INotebookViewModel } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
+import { getNotebookEditorFromEditorPane, INotebookEditor, INotebookEditorContribution, INotebookViewModel } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
 import { registerNotebookContribution } from 'vs/workbench/contrib/notebook/browser/notebookEditorExtensions';
 import { CodeCellViewModel } from 'vs/workbench/contrib/notebook/browser/viewModel/codeCellViewModel';
-import { INotebookCellDiagnosticsService } from 'vs/workbench/contrib/notebook/common/notebookCellDiagnosticsService';
 import { INotebookCellStatusBarItem, CellStatusbarAlignment } from 'vs/workbench/contrib/notebook/common/notebookCommon';
+import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
 
 export class DiagnosticCellStatusBarContrib extends Disposable implements INotebookEditorContribution {
 	static id: string = 'workbench.notebook.statusBar.diagtnostic';
@@ -40,17 +41,21 @@ class DiagnosticCellStatusBarItem extends Disposable {
 		private readonly _notebookViewModel: INotebookViewModel,
 		private readonly cell: CodeCellViewModel,
 		@IKeybindingService private readonly keybindingService: IKeybindingService,
-		@INotebookCellDiagnosticsService private readonly cellDiagnostics: INotebookCellDiagnosticsService
+		@IEditorService private readonly editorService: IEditorService
 	) {
 		super();
 		this._update();
-		this._register(this.cellDiagnostics.onDidDiagnosticsChange(() => this._update()));
+		const editor = getNotebookEditorFromEditorPane(this.editorService.activeEditorPane);
+		const diagnostics = editor?.getContribution<CellDiagnostics>(CellDiagnostics.ID);
+		if (diagnostics) {
+			this._register(diagnostics.onDidDiagnosticsChange(() => this._update()));
+		}
 	}
 
 	private async _update() {
 		let item: INotebookCellStatusBarItem | undefined;
 
-		if (!!this.cellDiagnostics.getCellExecutionError(this.cell.uri)) {
+		if (!!this.cell.excecutionError.get()) {
 			const keybinding = this.keybindingService.lookupKeybinding(OPEN_CELL_FAILURE_ACTIONS_COMMAND_ID)?.getLabel();
 			const tooltip = localize('notebook.cell.status.diagnostic', "Quick Actions {0}", `(${keybinding})`);
