+++ src/vs/workbench/contrib/notebook/browser/contrib/cellDiagnostics/cellDiagnostics.ts
@@ -3,151 +3,6 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { Disposable, IDisposable, toDisposable } from 'vs/base/common/lifecycle';
-import { IMarkerData, IMarkerService } from 'vs/platform/markers/common/markers';
-import { IRange } from 'vs/editor/common/core/range';
-import { ICellExecutionError, ICellExecutionStateChangedEvent, IExecutionStateChangedEvent, INotebookExecutionStateService, NotebookExecutionType } from 'vs/workbench/contrib/notebook/common/notebookExecutionStateService';
-import { IInlineChatService } from 'vs/workbench/contrib/inlineChat/common/inlineChat';
-import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { CellKind, NotebookSetting } from 'vs/workbench/contrib/notebook/common/notebookCommon';
-import { INotebookEditor, INotebookEditorContribution } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
-import { registerNotebookContribution } from 'vs/workbench/contrib/notebook/browser/notebookEditorExtensions';
-import { INotebookCellDiagnosticsService } from 'vs/workbench/contrib/notebook/common/notebookCellDiagnosticsService';
-import { Iterable } from 'vs/base/common/iterator';
-import { URI } from 'vs/base/common/uri';
-import { ResourceMap } from 'vs/base/common/map';
-
-type CellDiagnostic = {
-	cellUri: URI;
-	error: ICellExecutionError;
-	disposables: IDisposable[];
-};
-
-export class CellDiagnostics extends Disposable implements INotebookEditorContribution {
-
-	static ID: string = 'workbench.notebook.cellDiagnostics';
-
-	private enabled = false;
-	private listening = false;
-	private diagnosticsByHandle: Map<number, CellDiagnostic> = new Map();
-	private errors: ResourceMap<ICellExecutionError> = new ResourceMap();
-
-
-	constructor(
-		private readonly notebookEditor: INotebookEditor,
-		@INotebookCellDiagnosticsService private readonly notebookCellDiagnosticsService: INotebookCellDiagnosticsService,
-		@INotebookExecutionStateService private readonly notebookExecutionStateService: INotebookExecutionStateService,
-		@IMarkerService private readonly markerService: IMarkerService,
-		@IInlineChatService private readonly inlineChatService: IInlineChatService,
-		@IConfigurationService private readonly configurationService: IConfigurationService
-	) {
-		super();
-
-		this.updateEnabled();
-
-		this._register(inlineChatService.onDidChangeProviders(() => this.updateEnabled()));
-		this._register(configurationService.onDidChangeConfiguration((e) => {
-			if (e.affectsConfiguration(NotebookSetting.cellFailureDiagnostics)) {
-				this.updateEnabled();
-			}
-		}));
-	}
-
-	public getCellError(cellUri: URI): ICellExecutionError | undefined {
-		return this.errors.get(cellUri);
-	}
-
-	private updateEnabled() {
-		const settingEnabled = this.configurationService.getValue(NotebookSetting.cellFailureDiagnostics);
-		if (this.enabled && (!settingEnabled || Iterable.isEmpty(this.inlineChatService.getAllProvider()))) {
-			this.enabled = false;
-			this.clearAll();
-		} else if (!this.enabled && settingEnabled && !Iterable.isEmpty(this.inlineChatService.getAllProvider())) {
-			this.enabled = true;
-			if (!this.listening) {
-				this.listening = true;
-				this._register(this.notebookExecutionStateService.onDidChangeExecution((e) => this.handleChangeExecutionState(e)));
-			}
-		}
-	}
-
-	private handleChangeExecutionState(e: ICellExecutionStateChangedEvent | IExecutionStateChangedEvent) {
-		const notebookUri = this.notebookEditor.textModel?.uri;
-		if (this.enabled && e.type === NotebookExecutionType.cell && notebookUri && e.affectsNotebook(notebookUri)) {
-			if (!!e.changed) {
-				// cell is running
-				this.clear(e.cellHandle);
-			} else {
-				this.setDiagnostics(e.cellHandle);
-			}
-		}
-	}
-
-	private clearAll() {
-		for (const handle of this.diagnosticsByHandle.keys()) {
-			this.clear(handle);
-		}
-	}
-
-	public clear(cellHandle: number) {
-		const diagnostic = this.diagnosticsByHandle.get(cellHandle);
-		if (diagnostic) {
-			for (const disposable of diagnostic.disposables) {
-				disposable.dispose();
-			}
-			this.diagnosticsByHandle.delete(cellHandle);
-			this.errors.delete(diagnostic.cellUri);
-		}
-	}
-
-	private setDiagnostics(cellHandle: number) {
-		if (this.diagnosticsByHandle.has(cellHandle)) {
-			// multiple diagnostics per cell not supported for now
-			return;
-		}
-
-		const cell = this.notebookEditor.getCellByHandle(cellHandle);
-		if (!cell || cell.cellKind !== CellKind.Code) {
-			return;
-		}
-
-		const metadata = cell.model.internalMetadata;
-		if (!metadata.lastRunSuccess && metadata?.error?.location) {
-			const disposables: IDisposable[] = [];
-			const marker = this.createMarkerData(metadata.error.message, metadata.error.location);
-			this.markerService.changeOne(CellDiagnostics.ID, cell.uri, [marker]);
-			disposables.push(toDisposable(() => this.markerService.changeOne(CellDiagnostics.ID, cell.uri, [])));
-			disposables.push(this.notebookCellDiagnosticsService.registerCellExecutionError(cell.uri, metadata.error));
-			disposables.push(cell.model.onDidChangeOutputs(() => {
-				if (cell.model.outputs.length === 0) {
-					this.clear(cellHandle);
-				}
-			}));
-			disposables.push(cell.model.onDidChangeContent(() => {
-				this.clear(cellHandle);
-			}));
-			this.diagnosticsByHandle.set(cellHandle, { cellUri: cell.uri, error: metadata.error, disposables });
-			this.errors.set(cell.uri, metadata.error);
-		}
-	}
-
-	private createMarkerData(message: string, location: IRange): IMarkerData {
-		return {
-			severity: 8,
-			message: message,
-			startLineNumber: location.startLineNumber + 1,
-			startColumn: location.startColumn + 1,
-			endLineNumber: location.endLineNumber + 1,
-			endColumn: location.endColumn + 1,
-			source: 'Cell Execution Error'
-		};
-	}
-
-	override dispose() {
-		super.dispose();
-		this.clearAll();
-	}
-
-}
-
-registerNotebookContribution(CellDiagnostics.ID, CellDiagnostics);
+import 'vs/workbench/contrib/notebook/browser/contrib/cellDiagnostics/cellDiagnosticEditorContrib';
+import 'vs/workbench/contrib/notebook/browser/contrib/cellDiagnostics/cellDiagnosticsActions';
+import 'vs/workbench/contrib/notebook/browser/contrib/cellDiagnostics/diagnosticCellStatusBarContrib';
