+++ src/vs/workbench/contrib/notebook/browser/contrib/cellStatusBar/executionStatusBarItemController.ts
@@ -19,10 +19,6 @@ import { CellStatusbarAlignment, INotebookCellStatusBarItem, NotebookCellExecuti
 import { INotebookCellExecution, INotebookExecutionStateService, NotebookExecutionType } from 'vs/workbench/contrib/notebook/common/notebookExecutionStateService';
 import { INotebookService } from 'vs/workbench/contrib/notebook/common/notebookService';
 import { IMarkdownString } from 'vs/base/common/htmlContent';
-import { CodeCellViewModel } from 'vs/workbench/contrib/notebook/browser/viewModel/codeCellViewModel';
-import { OPEN_CELL_FAILURE_ACTIONS_COMMAND_ID } from 'vs/workbench/contrib/notebook/browser/contrib/cellCommands/cellCommands';
-import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
-import { INotebookCellDiagnosticsService } from 'vs/workbench/contrib/notebook/common/notebookCellDiagnosticsService';
 
 export function formatCellDuration(duration: number, showMilliseconds: boolean = true): string {
 	if (showMilliseconds && duration < 1000) {
@@ -335,63 +331,5 @@ class TimerCellStatusBarItem extends Disposable {
 
 		this._deferredUpdate?.dispose();
 		this._notebookViewModel.deltaCellStatusBarItems(this._currentItemIds, [{ handle: this._cell.handle, items: [] }]);
-	}
-}
-
-export class DiagnosticCellStatusBarContrib extends Disposable implements INotebookEditorContribution {
-	static id: string = 'workbench.notebook.statusBar.diagtnostic';
-
-	constructor(
-		notebookEditor: INotebookEditor,
-		@IInstantiationService instantiationService: IInstantiationService
-	) {
-		super();
-		this._register(new NotebookStatusBarController(notebookEditor, (vm, cell) =>
-			cell instanceof CodeCellViewModel ?
-				instantiationService.createInstance(DiagnosticCellStatusBarItem, vm, cell) :
-				Disposable.None
-		));
-	}
-}
-registerNotebookContribution(DiagnosticCellStatusBarContrib.id, DiagnosticCellStatusBarContrib);
-
-
-class DiagnosticCellStatusBarItem extends Disposable {
-	private _currentItemIds: string[] = [];
-
-	constructor(
-		private readonly _notebookViewModel: INotebookViewModel,
-		private readonly cell: CodeCellViewModel,
-		@IKeybindingService private readonly keybindingService: IKeybindingService,
-		@INotebookCellDiagnosticsService private readonly cellDiagnostics: INotebookCellDiagnosticsService
-	) {
-		super();
-		this._update();
-		this._register(this.cellDiagnostics.onDidDiagnosticsChange(() => this._update()));
-	}
-
-	private async _update() {
-		let item: INotebookCellStatusBarItem | undefined;
-
-		if (!!this.cellDiagnostics.getCellExecutionError(this.cell.uri)) {
-			const keybinding = this.keybindingService.lookupKeybinding(OPEN_CELL_FAILURE_ACTIONS_COMMAND_ID)?.getLabel();
-			const tooltip = localize('notebook.cell.status.diagnostic', "Quick Actions {0}", `(${keybinding})`);
-
-			item = {
-				text: `$(sparkle)`,
-				tooltip,
-				alignment: CellStatusbarAlignment.Left,
-				command: OPEN_CELL_FAILURE_ACTIONS_COMMAND_ID,
-				priority: Number.MAX_SAFE_INTEGER - 1
-			};
-		}
-
-		const items = item ? [item] : [];
-		this._currentItemIds = this._notebookViewModel.deltaCellStatusBarItems(this._currentItemIds, [{ handle: this.cell.handle, items }]);
-	}
-
-	override dispose() {
-		super.dispose();
-		this._notebookViewModel.deltaCellStatusBarItems(this._currentItemIds, [{ handle: this.cell.handle, items: [] }]);
 	}
 }
